Huge Fix
Bug fix where there was a problem when users abused the playing and stop functionality of youtube player. This made it go in a never ending loop of stop and play. It also sended messages non stop to the pusher server. The fix was adding an "allowance" of state changes of the player. This allowance resets every 5 seconds.