Few updates
Changed the vue chat to expand in medium sized screens. Changed the blade div to expand with jquery. Used vue to move the content of the messages like the thumbnails from center, to left/right. Also made sure the email address is not passed at the beggining to the probs by changing the select from roomcontroller show function.